Added 6.24 source

+++ clipssrc/prcdrfun.h
@@ -1,7 +1,7 @@
    /*******************************************************/
    /*      "C" Language Integrated Production System      */
    /*                                                     */
-   /*             CLIPS Version 6.30  08/16/14            */
+   /*             CLIPS Version 6.24  06/05/06            */
    /*                                                     */
    /*          PROCEDURAL FUNCTIONS HEADER FILE           */
    /*******************************************************/
@@ -11,26 +11,13 @@
 /*                                                           */
 /* Principal Programmer(s):                                  */
 /*      Gary D. Riley                                        */
-/*      Brian L. Dantes                                      */
+/*      Brian L. Donnell                                     */
 /*                                                           */
 /* Contributing Programmer(s):                               */
 /*                                                           */
 /* Revision History:                                         */
 /*                                                           */
-/*      6.23: Correction for FalseSymbol/TrueSymbol. DR0859  */
-/*                                                           */
-/*            Changed name of variable exp to theExp         */
-/*            because of Unix compiler warnings of shadowed  */
-/*            definitions.                                   */
-/*                                                           */
 /*      6.24: Renamed BOOLEAN macro type to intBool.         */
-/*                                                           */
-/*      6.30: Local variables set with the bind function     */
-/*            persist until a reset/clear command is issued. */
-/*                                                           */
-/*            Changed garbage collection algorithm.          */
-/*                                                           */
-/*            Support for long long integers.                */
 /*                                                           */
 /*************************************************************/
 
@@ -54,14 +41,14 @@
 
 typedef struct loopCounterStack
   {
-   long long loopCounter;
+   long loopCounter;
    struct loopCounterStack *nxt;
   } LOOP_COUNTER_STACK;
 
 #define PRCDRFUN_DATA 13
 
 struct procedureFunctionData
-  { 
+  {
    int ReturnFlag;
    int BreakFlag;
    LOOP_COUNTER_STACK *LoopCounterStack;
@@ -73,7 +60,7 @@ struct procedureFunctionData
    LOCALE void                           ProceduralFunctionDefinitions(void *);
    LOCALE void                           WhileFunction(void *,DATA_OBJECT_PTR);
    LOCALE void                           LoopForCountFunction(void *,DATA_OBJECT_PTR);
-   LOCALE long long                      GetLoopCount(void *);
+   LOCALE long                           GetLoopCount(void *);
    LOCALE void                           IfFunction(void *,DATA_OBJECT_PTR);
    LOCALE void                           BindFunction(void *,DATA_OBJECT_PTR);
    LOCALE void                           PrognFunction(void *,DATA_OBJECT_PTR);
@@ -83,7 +70,7 @@ struct procedureFunctionData
    LOCALE intBool                        GetBoundVariable(void *,struct dataObject *,struct symbolHashNode *);
    LOCALE void                           FlushBindList(void *);
 
-#endif /* _H_prcdrfun */
+#endif
